vendor docopt-ng
https://github.com/jazzband/docopt-ng/commit/fab01f5f19c9c7ca30f05e79a06d79bd066b475a

Docopt-ng is not backwards-compatible with Docopt. The change that
affects peru is that the `help` kwarg has been renamed to
`default_help`, but there are other changes that will affect other
projects. If we install docopt-ng as a dependency on e.g. Arch, it's
likely to break other projects that `import docopt` expecting the old
API. It seems like the best way to avoid trouble here is to vendor it.